New HomeMenu feature, WIP updater menu
Home menu items can not get a flag for if they are displayed in the list. If no, they will not be drawn in the top bar. This was added for the new update menu which is currently unfinished but exists in this commit. Also set up update silencing system